pyproject.toml
[project]
name = "business-assistant-contacts-plugin"
version = "0.1.0"
description = "Google Contacts plugin for Business Assistant v2"
requires-python = ">=3.11,<3.13"
dependencies = [
    "business-assistant-google-auth",
    "business-assistant-v2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/business_assistant_contacts"]

[tool.uv.sources]
business-assistant-v2 = { path = "D:/GIT/BenjaminKobjolke/business-assistant-v2", editable = true }
business-assistant-google-auth = { path = "D:/GIT/BenjaminKobjolke/business-assistant-google-auth", editable = true }

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
    "mypy>=1.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]

[tool.mypy]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
